feat: implement login mapping and enhance entity classes; refactor tests for improved setup
