+++ lib/domin/entities/login_entity.dart
@@ -1,5 +1,10 @@
-class LoginEntity {
+import 'package:equatable/equatable.dart';
+
+class LoginEntity extends Equatable {
   final String userToken;
   final String message;
-  LoginEntity({required this.userToken, required this.message});
+  const LoginEntity({required this.userToken, required this.message});
+
+  @override
+  List<Object?> get props => [userToken, message];
 }
